+++ kubejs/server_scripts/src/aileron/elytra.js
@@ -1,0 +1,10 @@
+ServerEvents.recipes(event => {
+    event.shaped('minecraft:elytra', [
+        'PPP',
+        'PGP',
+        'PPP'
+    ], {
+        P: 'minecraft:phantom_membrane',
+        G: 'immersiveengineering:glider'
+    });
+})